Edit source zipping script

+++ util/sourcebuilder.sh
@@ -1,5 +1,11 @@
 #!/bin/sh
 
+if [[ $# -eq 0 ]] ; then
+    echo Provide a version number e.g. 1.0.0
+    exit 1
+fi
+
+
 # Remove previous data if any
 echo Setting up...
 rm -rf raosource raosource.zip
@@ -18,6 +24,6 @@ cp $(ls -pa | grep -v / | grep -v md | grep -v txt) util/raosource # Copies all 
 # Zip the source
 echo Zipping...
 cd util
-tar.exe acf raosource.zip -C raosource .
+tar.exe acf raosource_v"$1".zip -C raosource .
 rm -rf raosource
 echo Complete.
